+++ JobInterview/FactorQuestion/Factor.h
@@ -13,7 +13,7 @@ public:
     static std::list<ULL> CreateFactors(ULL maxValue);
 private:
     static std::list<ULL> CreateFactorsRecursive(ULL value, ULL maxValue, int lastLevel);
-    static void Factor::CreateFactorsRecursiveCall(ULL value, ULL currentLevel, ULL maxValue, std::list<ULL> *factorList, std::future<std::list<ULL>> *futureFactorList);
+    static void Factor::CreateFactorsRecursiveCall(ULL value, int currentLevel, ULL maxValue, std::list<ULL> *factorList, std::future<std::list<ULL>> *futureFactorList);
 
     static int m_runningThreadCount;
     static std::mutex m_threadMutex;
